Added (meta) data loading & combining to excerpts
